requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0

# AI and LLM
openai>=1.12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
typing-extensions>=4.9.0
